Refactor tests to make it easier to add more example functions
This makes it so that it is possible to add a new example while only
modifying one location in the code. Add a new class / function / exception
in test_integration, indicate the expected output with the @expected decorator.
More efficient than having to modify three different files like we had to do
before.
