--- tests/roots/test-dummy/index.rst
@@ -1,59 +1,0 @@
-Dummy Module
-============
-
-.. autoclass:: dummy_module.Class
-   :members:
-   :undoc-members:
-   :private-members:
-   :special-members: __magic_custom_method__
-
-.. autoexception:: dummy_module.DummyException
-   :members:
-   :undoc-members:
-
-.. autofunction:: dummy_module.function
-
-.. autofunction:: dummy_module.function_with_escaped_default
-
-.. autofunction:: dummy_module.function_with_unresolvable_annotation
-
-.. autofunction:: dummy_module.function_with_typehint_comment
-
-.. autofunction:: dummy_module.function_with_starred_documentation_param_names
-
-.. autoclass:: dummy_module.ClassWithTypehints
-   :members:
-
-.. autofunction:: dummy_module.function_with_typehint_comment_not_inline
-
-.. autoclass:: dummy_module.ClassWithTypehintsNotInline
-   :members:
-
-.. autofunction:: dummy_module.undocumented_function
-
-.. autoclass:: dummy_module.DataClass
-    :undoc-members:
-    :special-members: __init__
-
-.. autodecorator:: dummy_module.Decorator
-
-.. autofunction:: dummy_module.mocked_import
-
-.. autofunction:: dummy_module.func_with_examples
-
-.. autofunction:: dummy_module.func_with_overload
-
-.. autoclass:: dummy_module.TestClassAttributeDocs
-   :members:
-
-.. autofunction:: dummy_module.func_with_examples_and_returns_after
-
-.. autofunction:: dummy_module.func_with_parameters_and_stuff_after
-
-.. autofunction:: dummy_module.func_with_rtype_in_weird_spot
-
-.. autofunction:: dummy_module.empty_line_between_parameters
-
-.. autofunction:: dummy_module.func_with_code_block
-
-.. autofunction:: dummy_module.func_with_definition_list
