+++ whitelist.txt
@@ -57,6 +57,7 @@ rootdir
 rst
 rtype
 runtime
+setitem
 sig
 signode
 skipif
